Hieu

+++ src/main/java/com/healthy/backend/dto/student/StudentResponse.java
@@ -1,16 +1,18 @@
 package com.healthy.backend.dto.student;
 
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.List;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.healthy.backend.dto.survey.SurveyResultsResponse;
 import com.healthy.backend.dto.user.UsersResponse;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @Builder
@@ -40,11 +42,11 @@ public class StudentResponse {
     @Schema(example = "Male")
     private String gender;
     @Schema(example = "0")
-    private Integer depressionScore;
+    private BigDecimal depressionScore;
     @Schema(example = "0")
-    private Integer anxietyScore;
+    private BigDecimal anxietyScore;
     @Schema(example = "0")
-    private Integer stressScore;
+    private BigDecimal stressScore;
     @Schema(examples = {"General", "Behavior", "Academic", "Emotional"} )
     private List<SurveyResultsResponse> surveyResults;
     @Schema(example = "2023-01-01")
